requirements.txt
discord.py-self
aiohttp
requests
protobuf<5
python-dotenv